Integrate Stat into Client and Index
